Add sublists to initial lists and to new item
Add sublists to initial lists and to new top-level items created with
the `Add` button.

It should be easier to add sublist items to top-level items this way; I
don’t have to check if a top-level item already contains an `ul` element
and can just add an `li` item to it straight away.