Limit the strike out text
Limit the strike out text to the top-level item by wrapping the text in
a `span` with attribute `tp-text` (“top-level text”) and changing the
CSS selector appropriately.

Now I need to make sure that second-level items also get a strike out
when they are clicked on, since this change breaks that behavior.