Treat non-visible items as `checked`
It makes sense that closed subitems don’t count towards the completion
of the top-level item/goal.  If you created a subitem/subgoal and then
closed it without checking it first, it would be impossible to close the
top-level item by closing all the visible subitems; since the closed
subitem still existed (it was just non-visible) but wasn’t checked, not
all children of the top-level item would be `checked`.  So add another
condition, making a disjunction: a node is considered to be done if
either (1) it is `checked`, *or* (2) it is non-visible.

I also tried to add the `checked` class to subitems when they are
closed (see `setOnclickForClose(_)`), but that didn’t work for whatever
reason.